Fix ROUND type error that silently killed every later migration
ROUND(double precision, integer) does not exist in Postgres, and
purchase_order_items.amount is double precision. Postgres resolves the
function at plan time, so this threw even though the =480 guard matched no
rows. Cast the expression to numeric.

initDB() swallows its own errors (index.js:107 logs and moves on so the
server still boots when the DB is briefly unreachable), so this failed
silently on every deploy since da3f75c on 3 Aug — and took everything
after it in initDB with it. Found while a new reconciliation block at the
end of the file refused to apply.

Casualties were smaller than feared: the tables and constraints in the
dead zone are all IF-NOT-EXISTS or already matched production. The one
real gap was the RLS pass, so bank_accounts, petty_cash_attachments,
purchase_debit_notes and purchase_item_rejection_photos sat without row
level security — the only tables in the schema exposed to Supabase's
auto-generated anon REST API. Now enabled.

+++ server/src/db/index.js
@@ -1307,7 +1307,7 @@ async function initDB(retries = 20, delayMs = 10000) {
       // makes this run once; shares never equal exactly 480 with multiple items).
       await pool.query(`
         UPDATE purchase_order_items poi
-           SET receive_transport_cost = ROUND(480.0 * poi.amount / NULLIF(t.total, 0), 2)
+           SET receive_transport_cost = ROUND((480.0 * poi.amount / NULLIF(t.total, 0))::numeric, 2)
           FROM purchase_orders po,
                (SELECT SUM(poi2.amount) AS total
                   FROM purchase_order_items poi2 JOIN purchase_orders po2 ON po2.id = poi2.po_id
